Makes evaluator field an object and refactors asset reading logic.
The logic to read asset and adjust endianness has been moved into an AssetUtils class.

The field outputMeanStdDev in the evaluator config can be used to specify
an asset containing means and standard deviations for a model's outputs, which
can then be used for output de-normalization. This is needed for models
such as TTS where the validation requires denormalized outputs.

Bug: 116307951
Bug: 116315300

Test: Covered by existing test:
com.example.android.nn.benchmark.NNTest

Change-Id: Ic8ccefbf2bac1fb8c542462536e70e1e660e82d3