Generate results summary HTML page.
Added ability to run selected models/datasets/backends
for a 'scoring' run. It produces a CSV file with
detailed results of benchmark run. It then can be
processed into a single, self-contained HTML page with
benchmark results. This result page is intended for internal
and 3rd parties usage and will form a basis for a single score
calculation.

Scoring run can be started by running:
./test/mlts/benchmark/build_and_run_benchmark.sh

Bug: 115720411
Test: build_and_run_benchmark.sh
Change-Id: I36e4f54cd19d2737ab449ea25f01ee696ea6dcda