Separates UnsupportedSdkException from BenchmarkException.
When running the app, we want to display an error message for both cases.
When running test, we want the test to fail for the latter and not the former.

Bug: 117759123

Test: Ran the app and the test com.android.nn.benchmark.app.NNTest

Change-Id: Ic43d5b8928a73b7cbef5dc113e9928f246f7027e
Merged-In: Ic43d5b8928a73b7cbef5dc113e9928f246f7027e
(cherry picked from commit 71443e37b62aacc66827257536dfa6d75885edc2)