Cause test failure in case of TFLite CPU Fallback
Add checks to result of nnapi_delegate.GetNnapiErrno() in order
to recognise failures in NNAPI even if the execution of the TFLite model
is then completed by the TFLite CPU Fallback.
Causing failure if a model cannot be loaded or delegated to NNAPI.
Fixed issue causing an early failure in Processor's execution not being
recognized and notification progress message not being printed out.

Test: build_and_run_benchmark.sh parallel-inference-stress
Change-Id: I0fe5745855809b49673f460c858c7b1ca6fe8dca