Refactor test code to prepare for multi process tests
This is a preparatory change to refactor existing code in preparation
for the addition of multi-process tests (actually is a split of the work
done in r.android.com/1300134).

Extracting the test status management logic from NNParallelTestActivity.
Extracting logic to find models supported by accelerators and to
validate if a driver crashed in an external process.

Test: build_and_run_benchmark.sh parallel-inference-stress
Change-Id: I358a33fd8e8bbdcf4d2154f7947f7bb9273a4aee