Add test for memory mapped models
Add flag to control how TFLite model is loaded, memory mapped to the
underlying model file or from the program memory.
Create a separate test to validate the behaviour of drivers when dealing
with memory mapped models and sets the default behaviour for other tests
to the safest option, loaded from program memory.
Add MMAP option to the crash test app.

Test: build_and_run_benchmark.sh memory-mapped-model-load-stress
Change-Id: Ia1797c47de64137ddb811b6a170a81eeaba8f433