Restore multi-process stress tests
Add a new set of inference, model load and early termination tests
running on multiple processes. Early termination tests are included in
the inference and model load tests with difference client process
mortality rates.
This CL fixes problem the problems in the first commit breaking the
build when merged with rvc-dev.

Test: build_and_run_benchmark multi-process-inference-stress

This reverts commit c49b55e3594cbb546c7eaf6d641ebf5c2b0b5dc7.

Change-Id: I7ab1b31b8e0dfbf7b1ac653b7f065d9d205680e4