Restore multi-process stress tests
Add a new set of inference, model load and early termination tests
running on multiple processes. Early termination tests are included in
the inference and model load tests with difference client process
mortality rates.
This CL fixes problem the problems in the first commit breaking the
build when merged with rvc-dev.

Test: build_and_run_benchmark multi-process-inference-stress
Bug: 152733986
Change-Id: I5004add7123cef0339404096dc48b1c76779020e