Add compilation benchmark to MLTS.
The compilation benchmark is enabled in NNScoringTest. Four metrics
may be reported:
- The compilation latency without caching
- The compilation latency upon cache miss
- The compilation latency upon cache hit
- The total size of cache files

None of the metrics will be reported with TFLite_CPU backend.

The benchmark will try to detect if compilation caching is supported by
checking the cache file size, and will not benchmark cache miss / cache
hit if compilation caching is not supported.

Bug: 154013612
Test: ./build_and_run_benchmark.sh scoring
Change-Id: Ia1de97732ec65f8b7d109f07a6bd962137cb865c