Fix failure in scoring test caused by timeout
Manually merging https://r.android.com/1391256 to solve conflicts.

Fix failures "IllegalStateException: Failed to evaluate complete input
set" in scoring tests.
Tests couldn't complete because they were running over the configured
timeout.

Refactored the code to clarify the purpose of the timeout parameter that
was used in parts of the code as a minimum running time.

Test: build_and_run_benchmark.sh scoring
Bug: 159782692
Change-Id: I6c9d48330f599256006b3dda37078df5a8288469